Radius Upgrade Implemented, Zombie Creation Changed
Radius upgrade was made but it doesnt seem to update the variable dynamically so it doesnt actually increase in size when the upgrade is completed. Changed triggers so when a civillian dies they just spawn a zombie where they died, using regions to create zombies from corpses when another zombie walks over it was buggy and weird.